Replace VictoryBar component
