+++ packages/victory-bar/src/index.js
@@ -1,4 +1,4 @@
-export { default as VictoryBar } from "./victory-bar-duplicate";
+export { default as VictoryBar } from "./victory-bar";
 export { default as Bar } from "./bar";
 export { getBarPosition } from "./helper-methods";
 export {
